pyproject.toml
[tool.poetry]
name = "b2b-marketplace-backend"
version = "0.1.0"
description = ""
authors = ["b2b marketplace team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"
djoser = "^2.2.0"
drf-spectacular = "^0.26.4"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
black = "^23.7.0"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
exclude = '.+/migrations/.+\.py$'

[tool.pytest.ini_options]
pythonpath = ["b2b-marketplace-backend"]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
addopts = "-vv -p no:cacheprovider --disable-warnings"
norecursedirs = [".*"]
